// src/generated/server/alsaglobal/research/v1/service_server.ts
export interface PaginationResponse {
  nextCursor: string;
  totalCount: number;
}

export interface ListTrendingReposRequest {
  pageSize: number;
  cursor: string;
  language: string;
  period: string;
}

export interface ListTrendingReposResponse {
  repos: GithubRepo[];
  pagination?: PaginationResponse;
}

export interface GithubRepo {
  fullName: string;
  description: string;
  language: string;
  stars: number;
  starsToday: number;
  forks: number;
  url: string;
}
// ...
export interface ServerContext {
  request: Request;
  pathParams: Record<string, string>;
  headers: Record<string, string>;
}